+++ iOS/PrivateFrameworks/GeoServices.framework/GEOResourceManifestServerProxy-Protocol.h
@@ -16,7 +16,7 @@
 - (oneway void)resetActiveTileGroup;
 - (void)setActiveTileGroupIdentifier:(NSNumber *)arg1 updateType:(long long)arg2 completionHandler:(void (^)(NSError *))arg3;
 - (oneway void)setActiveTileGroupIdentifier:(NSNumber *)arg1;
-- (unsigned long long)maximumZoomLevelForStyle:(int)arg1 scale:(int)arg2 outSize:(int *)arg3;
+- (unsigned int)maximumZoomLevelForStyle:(int)arg1 scale:(int)arg2 outSize:(int *)arg3;
 - (NSProgress *)updateProgress;
 - (void)deactivateResourceScenario:(int)arg1;
 - (void)activateResourceScenario:(int)arg1;
